Added support for Arduino Leonardo with ATmega32u4 and for older Arduino Serial devices with the ATmega8.
To flash the LED when they have stack overflows and/or malloc allocation
failures. That pretty much covers the field for Arduino AVR devices, as
they all have the "same" watchdog timer.

Increased the configured loop() function stack size to 256 bytes. This
is a reasonable compromise for RAM usage for 328p based devices.
Additional tasks can be configured with any stack size on creation.