Reverting __attribute__((weak)) function move.
The need to keep weak functions out of the file they are called in only
applies if a library is trying to replace the weak function call.

If individual files are linked together, this issue doesn't seem to
apply.

So that is better. Nothing to do, except make a library.